sync schema before calling transact
This lets you add :db/index as schema alterations without having to
deploy two sets of production changes and manually call synch-schema
between them. For the average user, it won't add any extra time to the
time for migrations.

This also lets the user pass a timeout for sync-schema in the norm map,
via setting :conformity.setting/sync-schema-timeout to a value in
milliseconds. This is passed directly to deref. If the timout is hit,
conformity will throw an exception.