Add tx-instant argument (#51)
* Add tx-instant argument

One use case is an initial import of existing data that has its own timestamps, where the timestamps are used as the `:db/txInstant` value. To enable the use of Conformity *before* the initial import, a custom `:db/txInstant` value (rather than the transactor's clock time) is necessary. This is because `:db/txInstant` cannot be set to a value that is older than any existing transaction.

If `tx-instant` argument is passed, Conformity will use it as `:db/txInstant` attribute of the transactions while updating database schema. If it is not passed, the current behaviour is maintained.

The `last-tx-instant` utility function can be used together with the `tx-instant` argument to keep the newest transaction time as it is.

* Replace cond->> with if

* Add Maintaining txInstant section to README